[ refactor ] index mods & commands by their name
I got fed up with the potential discrepancy between the field name
& the name stored inside of the command so here is a refactoring
indexing a command (and a modifier) by its name.

We need to change the definition of `Fields` a bit but that enables
us to define cute smart constructors that mean we can get rid of the
hacky castGen & friends using snoc lists.

Cf. the test case for an updated design.