[ refactor ] accumulating errors (#14)
Also fixing the totality issues detected by my recent changes to
the termination checker and renaming Data.Record.Ordered given that
we are not in fact relying on an order.

Desugaring of do blocks is not based on Monad anyways so we don't
have to provide an implementation if we are not happy about the fact
it's not lawful. So we define `(>>=)` for `Error` but not `Monad Error`.